requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
black==23.3.0
click==8.1.3
colorama==0.4.6
coverage==7.2.7
Django==4.2.2
django-bootstrap-v5==1.0.11
django-pwa==1.1.0
django-stdimage==6.0.1
exceptiongroup==1.1.3
flake8==6.0.0
iniconfig==2.0.0
mccabe==0.7.0
model-bakery==1.12.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
Pillow==10.0.0
platformdirs==3.6.0
pluggy==1.2.0
psycopg2-binary==2.9.6
pycodestyle==2.10.0
pyflakes==3.0.1
PyMySQL==1.1.0
pytest==7.4.0
pytest-django==4.5.2
python-dotenv==1.0.0
soupsieve==2.4.1
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.6.3
tzdata==2023.3
whitenoise==6.5.0
